requirements.txt
google-adk
deprecated
litellm
google-generativeai
python-dotenv
requests
Flask
Flask-CORS
SQLAlchemy
streamlit
requests
fastapi
uvicorn[standard]
deprecated
supabase

